Added an in-memory LRU Stored Requests cache.
