+++ stored_requests/caches/in_memory/lru.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"github.com/coocood/freecache"
 	"github.com/golang/glog"
+	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/stored_requests"
 )
 
@@ -14,10 +15,10 @@ import (
 // 2. The cache is too large. This will cause the least recently used items to be evicted.
 //
 // For no TTL, use ttlSeconds <= 0
-func NewLRUCache(sizeBytes int, ttlSeconds int) stored_requests.Cache {
+func NewLRUCache(cfg *config.InMemoryCache) stored_requests.Cache {
 	return &cache{
-		lru:        freecache.NewCache(sizeBytes),
-		ttlSeconds: ttlSeconds,
+		lru:        freecache.NewCache(cfg.Size),
+		ttlSeconds: cfg.TTL,
 	}
 }
