+++ stored_requests/caches/in_memory/lru_test.go
@@ -4,11 +4,15 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"github.com/prebid/prebid-server/config"
 	"testing"
 )
 
 func TestCacheMiss(t *testing.T) {
-	cache := NewLRUCache(512*1024, -1)
+	cache := NewLRUCache(&config.InMemoryCache{
+		Size: 512 * 1024,
+		TTL:  -1,
+	})
 	data := cache.GetRequests(context.Background(), []string{"unknown"})
 	if len(data) > 0 {
 		t.Errorf("An empty cache should not return any data on unknown IDs.")
@@ -16,7 +20,10 @@ func TestCacheMiss(t *testing.T) {
 }
 
 func TestCacheHit(t *testing.T) {
-	cache := NewLRUCache(512*1024, -1)
+	cache := NewLRUCache(&config.InMemoryCache{
+		Size: 512 * 1024,
+		TTL:  -1,
+	})
 	cache.SaveRequests(context.Background(), map[string]json.RawMessage{
 		"known": json.RawMessage(`{}`),
 	})
@@ -34,7 +41,10 @@ func TestCacheHit(t *testing.T) {
 }
 
 func TestCacheMixed(t *testing.T) {
-	cache := NewLRUCache(512*1024, -1)
+	cache := NewLRUCache(&config.InMemoryCache{
+		Size: 512 * 1024,
+		TTL:  -1,
+	})
 	cache.SaveRequests(context.Background(), map[string]json.RawMessage{
 		"known": json.RawMessage(`{}`),
 	})
